utils: fix up overload statements
It has never been 100% clear to me on the semantics of how
overload works, particularly with Literal parameters which are
optional (so have a default). The typing documentation has
improved so try to match the code to my current understanding
(I believe there may still be differences in how the different
Python tyoing engines deal with overloads).