data: overload typing tweaks
Remove the get_bounding_mask overload because it does not make sense:
there's no way to determine from the types what the return value
should be.

Improve the handling of True/False in overload statements when there's
a default value: rather than "object: Literal[False]" we say
"Literal[False] = False", which is a bit wordy but is supported in
pyright and mypy. There is an interesting change needed because we
end up having to say "Literal[True] = True" even though this can not
be the default value just to avoid the runtime from falling over with
a "non-default argument follows default argument" error.

Add explicit types to the runtime version of overloaded routines. This
no-longer seems to be needed but makes things a bit easier to check.